requirements.txt
Flask==3.1.1
python-dotenv==1.1.1
requests==2.32.4